Refactor thread functions to an own file
