+++ macke/threads.py
@@ -1,0 +1,42 @@
+"""
+All functions, that are executed in parallel threads
+"""
+
+from .Klee import execute_klee, execute_klee_targeted_search
+
+
+def thread_phase_one(
+        resultlist, functionname, symmains_bc, outdir,
+        flags, posixflags, posix4main):
+    """
+    This function is executed by the parallel processes in phase one
+    """
+    # Just run KLEE on it
+    try:
+        resultlist.append(execute_klee(
+            symmains_bc, functionname, outdir, flags, posixflags, posix4main))
+    except Exception as exc:
+        print()
+        print("A thread in phase one throws and exception")
+        print("The analyzed function was:", functionname)
+        print(exc)
+        print()
+
+
+def thread_phase_two(
+        resultlist, caller, callee, prepended_bc, outdir,
+        flags, posixflags, posix4main):
+    """
+    This function is executed by the parallel processes in phase two
+    """
+    # And run klee on it
+    try:
+        resultlist.append(execute_klee_targeted_search(
+            prepended_bc, caller, callee, outdir,
+            flags, posixflags, posix4main))
+    except Exception as exc:
+        print()
+        print("A thread in phase two throws and exception")
+        print("The analyzed caller/callee pair was:", caller, callee)
+        print(exc)
+        print()
